requirements.txt
beautifulsoup4==4.12.3
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
nlpia2_wikipedia==1.5.18
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
RapidFuzz==3.10.0
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
tomlkit==0.13.2
trove-classifiers==2024.10.16
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
zipp==3.20.2
